refactor(FeedVersionDetails): Extract MTC actions out
